Added tests for RoaringBitSet logic on varied size BitsetsNET
The RoaringBitSet uses two different types of containers, Array and Bitset.
The current tests only test bitsets of (approx) equal size so they do not expose issues with logical
operations done between an Array and Bitset container. These new tests will check that the logic
for an AND and an OR between a bitset and array container works as expected.