Fixed Cardinality calculation when performing BitsetContainer OR ArrayContainer
When the branchless OR functionality was added the assumption seems to have been made that
the shift right operator shifted in 0s (logical shift) but in reality it duplicates the most significant bit
(arithmetic shift). This means when we shift a negative long by 63 bits we get -1. The desired behavior
is to get a +1 in this case so by adding & 1 the -1 becomes +1. This could have been done by negating
the value after the shift too, but using the &1 more clearly shows the value can only be 0 or 1
(this was already intended to be the case).